refactor: changing file name from main to home
